# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Days Off App",
    "description": "Thanks to this application , the leave management system empowers your staff on . question such as How many days of paid vacation do i have left?  Your staff can check balances ,manage and track their time off.Managers can run reports,analyze absence trends,and find solutions to staffing issues in real time."
  },
  "servers": [
    {
      "url": "http://localhost:4500/"
    }
  ],
...
  "paths": {
...
    "/daysOff": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "daysoffs"
        ],
        "summary": "Create a new days off ",
        "description": "Create a new days off ",
        "operationId": "AddDaysOff",
        "requestBody": {
          "description": "Created daysOff object",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startDay": {
                    "type": "string",
                    "example": "2023-01-01"
                  },
                  "endDay": {
                    "type": "string",
                    "example": "2023-01-05"
                  },
                  "type": {
                    "type": "string",
                    "example": "Paid"
                  },
                  "justificationSick": {
                    "type": "string",
                    "example": "certificat medical"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Days"
                }
              }
            }
          },
          "400": {
            "description": "Bad request "
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "daysoffs"
        ],
        "summary": "Get the list of all days off",
        "description": "Get the list of all days off",
        "operationId": "ListDaysOff",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": true,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": true,
            "type": "integer"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Field to sort by",
            "required": true,
            "type": "string"
          },
          {
            "name": "createdAtBefore",
            "in": "query",
            "description": "Filter users created before this date (YYYY-MM-DD format)",
            "required": true,
            "type": "string"
          },
          {
            "name": "createdAtAfter",
            "in": "query",
            "description": "Filter users created after this date (YYYY-MM-DD format)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succussffully updated ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 1
                    },
                    "totalDaysOff": {
                      "type": "integer",
                      "example": 1
                    },
                    "daysOff": {
                      "type": "array",
                      "example": [
                        {
                          "decisionDirector": {
                            "status": true,
                            "justification": "string",
                            "userIdDir": "string"
                          },
                          "decisionManager": {
                            "status": false,
                            "justification": "string",
                            "userIdMan": "string"
                          },
                          "_id": "string",
                          "userId": "string",
                          "startDay": "string",
                          "endDay": "string",
                          "type": "string",
                          "statusReq": false,
                          "statusDecision": true,
                          "reqDayOff": 0
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "request not found !"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "daysoffs"
        ],
        "summary": "Delete all request",
        "description": "Delete all request",
        "operationId": "deleteAllRequest",
        "responses": {
          "200": {
            "description": " All daysOff are succussffully deleted"
          },
          "404": {
            "description": "Requests not found or you are disabled now!"
          },
          "401": {
            "description": "you can not remove all request!"
          }
        }
      }
    },
    "/daysOff/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "daysoffs"
        ],
        "summary": "Get days off whose identifier is known",
        "description": "Get days off whose identifier is known",
        "operationId": "GetDaysOff",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Days"
                }
              }
            }
          },
          "404": {
            "description": "request not found !"
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "daysoffs"
        ],
        "summary": "Update days off whose identifier is known",
        "description": "Update days off whose identifier is known",
        "operationId": "UpdateDaysOff",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The request that needs to updated",
            "required": true
          }
        ],
        "requestBody": {
          "description": "Request with new properties",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "startDay": {
                    "type": "string",
                    "example": "2023-01-18"
                  },
                  "endDay": {
                    "type": "string",
                    "example": "2023-01-22"
                  },
                  "type": {
                    "type": "string"
                  },
                  "justificationSick": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "succussffully updated ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "startDay": {
                      "type": "string",
                      "example": "2023-01-01"
                    },
                    "endDay": {
                      "type": "string",
                      "example": "2023-10-01"
                    },
                    "type": {
                      "type": "string",
                      "example": "Paid"
                    },
                    "justificationSick": {
                      "type": "string",
                      "example": "certificat medical"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "not found"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "daysoffs"
        ],
        "summary": "Delete request whose identifier is known",
        "description": "Delete request whose identifier is known",
        "operationId": "deleteRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The request that needs to be deleted",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "days off is succussffully deleted"
          },
          "400": {
            "description": "you can not remove this request!"
          },
          "404": {
            "description": "Request not found or you are disabled now!"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Days": {
        "required": [
          "userId",
          "startDay",
          "endDay",
          "type"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "string"
          },
          "startDay": {
            "type": "string",
            "example": "2023-01-18"
          },
          "endDay": {
            "type": "string",
            "example": "2023-01-22"
          },
          "type": {
            "type": "string",
            "enum": [
              "Paid",
              "Unpaid",
              "Sick"
            ]
          },
          "decisionManager": {
            "type": "object",
            "properties": {
              "userIdMan": {
                "type": "string"
              },
              "status": {
                "type": "boolean"
              },
              "justification": {
                "type": "string"
              }
            }
          },
          "decisionDirector": {
            "type": "object",
            "properties": {
              "userIdDir": {
                "type": "string"
              },
              "status": {
                "type": "boolean"
              },
              "justification": {
                "type": "string"
              }
            }
          },
          "statusReq": {
            "type": "boolean",
            "example": "true"
          },
          "statusDecision": {
            "type": "string",
            "example": "true"
          },
          "reqDayOff": {
            "type": "Number",
            "example": "0"
          },
          "justificationSick": {
            "type": "string",
            "example": "sick"
          }
        }
      }
    },
...
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    }
  }
}